backend/requirements.txt
faiss-cpu
langchain
langchain-community
transformers
torch
numpy
sentence-transformers
python-dotenv
pydantic
accelerate
einops
tqdm
flask
flask-cors